+++ src/main/kotlin/us/frollo/frollosdk/model/coredata/surveys/Survey.kt
@@ -16,6 +16,7 @@
 
 package us.frollo.frollosdk.model.coredata.surveys
 
+import com.google.gson.JsonObject
 import com.google.gson.annotations.SerializedName
 
 /**
@@ -31,6 +32,15 @@ data class Survey(
 
     /** Name of the survey */
     @SerializedName("name") val name: String?,
+
+    /** Generic metadata JSON field */
+    @SerializedName("metadata") val metadata: JsonObject?,
+
+    /** title of the survey */
+    @SerializedName("display_text") val displayText: String?,
+
+    /** icon url for survey */
+    @SerializedName("icon_url") val iconUrl: String?,
 
     /** List of all questions within the survey. */
     @SerializedName("questions") val questions: List<SurveyQuestion>